chore: move filter to general module
