+++ packages/shared/components/Proposals.svelte
@@ -1,9 +1,9 @@
 <script lang="typescript">
-    import { Text, ProposalCard } from 'shared/components'
+    import { Text, ProposalCard, Filter } from 'shared/components'
     import { localize } from '@core/i18n'
     import { FontWeight } from './enums'
-
     import { IProposal } from '@contexts/governance/interfaces'
+    import { proposalFilter } from '@contexts/governance'
 
     export let proposals: IProposal[] = []
 
@@ -15,6 +15,9 @@
         <Text fontSize="14" fontWeight={FontWeight.semibold}>
             {localize('views.governance.proposals.title')}
         </Text>
+        <div class="flex flex-row">
+            <Filter filterStore={proposalFilter} />
+        </div>
     </header-container>
     <ul class="grid grid-cols-2 auto-rows-min gap-6 flex-1 overflow-y-scroll">
         {#each sortedProposals as proposal}
