+++ packages/shared/components/filters/Filter.svelte
@@ -1,12 +1,13 @@
 <script lang="typescript">
     import { TogglableButton, FilterModal, FilterItem, Modal } from 'shared/components'
-    import type { Filter } from '@core/wallet/interfaces/filter/filter.interface'
+    import type { ActivityFilter, AssetFilter } from '@core/wallet/interfaces'
     import { deepCopy } from '@core/utils'
     import type { Writable } from 'svelte/store'
+    import { ProposalFilter } from '@contexts/governance'
 
-    export let filterStore: Writable<Filter>
+    export let filterStore: Writable<ActivityFilter | AssetFilter | ProposalFilter>
 
-    let filter: Filter = deepCopy($filterStore)
+    let filter: ActivityFilter | AssetFilter | ProposalFilter = deepCopy($filterStore)
 
     let filterActive = false
     let modal: Modal
