+++ packages/shared/components/filters/items/OrderFilterItem.svelte
@@ -1,9 +1,9 @@
 <script lang="typescript">
-    import type { OrderFilterUnit } from '@core/wallet/interfaces'
     import { Dropdown } from 'shared/components'
     import { localize } from '@core/i18n'
     import type { IDropdownChoice } from '@core/utils'
-    import { OrderOption } from '@core/wallet'
+    import { OrderFilterUnit } from '@core/utils/interfaces/filter'
+    import { OrderOption } from '@core/utils/enums/filters'
 
     export let filterUnit: OrderFilterUnit
