+++ packages/shared/components/filters/items/SelectionFilterItem.svelte
@@ -1,8 +1,8 @@
 <script lang="typescript">
-    import type { SelectionFilterUnit } from '@core/wallet/interfaces'
     import { Dropdown } from 'shared/components'
     import { localize } from '@core/i18n'
     import type { IDropdownChoice } from '@core/utils'
+    import { SelectionFilterUnit } from '@core/utils/interfaces/filter'
 
     export let filterUnit: SelectionFilterUnit
