+++ packages/shared/lib/contexts/governance/interfaces/index.ts
@@ -1,4 +1,5 @@
 export * from './organization.interface'
+export * from './proposal-filter.interface'
 export * from './proposal-state.interface'
 export * from './proposal.interface'
 export * from './proposals-details.interface'
