+++ packages/shared/lib/contexts/governance/interfaces/proposal-filter.interface.ts
@@ -1,0 +1,7 @@
+import { SelectionFilterUnit } from '@core/utils/interfaces/filter/filter-unit.interface'
+
+export interface ProposalFilter {
+    phase: SelectionFilterUnit
+    type: SelectionFilterUnit
+    participated: SelectionFilterUnit
+}
