+++ packages/shared/lib/contexts/governance/stores/index.ts
@@ -1,5 +1,6 @@
 export * from './has-to-revote.constant'
 export * from './participation-overview.store'
+export * from './proposal-filter.store'
 export * from './proposals-state.store'
 export * from './registered-event-ids.store'
 export * from './selected-proposal.store'
