+++ packages/shared/lib/core/utils/interfaces/filter/date-filter-input.interface.ts
@@ -1,4 +1,4 @@
-import { DateUnit } from '@core/wallet/enums'
+import { DateUnit } from '@core/utils/enums/filters'
 
 export type DateFilterInput = SingleDateFilterInput | RangeDateFilterInput | UnitDateFilterInput
