+++ packages/shared/lib/core/utils/interfaces/filter/index.ts
@@ -1,4 +1,3 @@
 export * from './date-filter-input.interface'
 export * from './filter-unit.interface'
-export * from './filter.interface'
 export * from './number-filter-input.interface'
