+++ packages/shared/lib/core/wallet/enums/index.ts
@@ -8,6 +8,5 @@ export * from './governance-action.enum'
 export * from './not-verified-status.enum'
 export * from './token-standard.enum'
 export * from './verified-status.enum'
-export * from './filters'
 export * from './inclusion-state.enum'
 export * from './irc27-version.enum'
