+++ packages/shared/lib/core/wallet/interfaces/asset-filter.interface.ts
@@ -1,0 +1,7 @@
+import { OrderFilterUnit, SelectionFilterUnit } from '@core/utils/interfaces/filter'
+
+export interface AssetFilter {
+    verificationStatus: SelectionFilterUnit
+    showHidden: SelectionFilterUnit
+    order: OrderFilterUnit
+}
