+++ packages/shared/lib/core/wallet/interfaces/index.ts
@@ -7,8 +7,8 @@ export * from './asset.interface'
 export * from './base-token-metadata.interface'
 export * from './processed-transaction.interface'
 export * from './claimed-activities.interface'
-export * from './filter/filter.interface'
-export * from './filter'
+export * from './activity-filter.interface'
+export * from './asset-filter.interface'
 export * from './participation.interface'
 export * from './hidden-activities.interface'
 export * from './irc30-metadata.interface'
