+++ packages/shared/lib/core/wallet/stores/selected-account-assets.store.ts
@@ -1,9 +1,10 @@
 import { selectedAccount } from '@core/account/stores/selected-account.store'
 import { marketCoinPrices } from '@core/market'
 import { activeProfileId } from '@core/profile'
+import { AssetOrderOption, BooleanFilterOption, OrderOption } from '@core/utils/enums/filters'
 import { derived, get, Readable, writable, Writable } from 'svelte/store'
 import { getAccountAssetsForSelectedAccount } from '../actions/getAccountAssetsForSelectedAccount'
-import { AssetOrderOption, BooleanFilterOption, NotVerifiedStatus, OrderOption, VerifiedStatus } from '../enums'
+import { NotVerifiedStatus, VerifiedStatus } from '../enums'
 import { AssetFilter, IAsset } from '../interfaces'
 import { IAccountAssets } from '../interfaces/account-assets.interface'
 import { persistedAssets } from './persisted-assets.store'
