+++ packages/shared/lib/core/wallet/utils/isVisibleAsset.ts
@@ -1,5 +1,5 @@
+import { BooleanFilterOption } from '@core/utils/enums/filters'
 import { get } from 'svelte/store'
-import { BooleanFilterOption } from '../enums'
 import { IPersistedAsset } from '../interfaces/persisted-asset.interface'
 import { assetFilter } from '../stores'
